+++ start/src/App.vue
@@ -9,6 +9,17 @@
 </script>
 
 <template>
+  <f-link href="" type="default">默认链接</f-link>
+  <f-link href="" type="primary">主要链接</f-link>
+  <f-link href="" type="success">成功链接</f-link>
+  <f-link href="" type="danger">危险链接</f-link>
+  <f-link href="" type="warning">警告链接</f-link>
+
+  <br />
+  <f-link disabled href="" type="primary">不许点我</f-link>
+  <f-link disabled href="" state="line" type="success">不许点我</f-link>
+  <f-link disabled href="" state="bag" type="warning">不许点我</f-link>
+
   <!-- <f-menu>
     <template #before>
       <f-avatar